Made few TODOs into NOTE and ALT, for general notes about the code and alternative solutions
Moved Extensions into separate namespace Helpers.Extensions

Changed resources to floating point values instead of integer values, to enable continuous resource collection
It can still be used as integer up to pretty big numbers

Changed saving and loading resources accordingly

Added spawning units, buildings and projectiles with some initial rotation, added saving and loading these rotations

Changed back to single building per tile to make getting height at points easier

+++ MHUrho/MHUrho/Logic/IPlayer.cs
@@ -22,7 +22,7 @@ namespace MHUrho.Logic {
 
 		IReadOnlyList<IBuilding> GetBuildingsOfType(BuildingType buildingType);
 
-		int GetResourceAmount(ResourceType resourceType);
+		double GetResourceAmount(ResourceType resourceType);
 
 		IEnumerable<IPlayer> GetEnemyPlayers();
 
@@ -38,6 +38,6 @@ namespace MHUrho.Logic {
 
 		bool RemoveBuilding(IBuilding building);
 
-		void ChangeResourceAmount(ResourceType resourceType, int change);
+		void ChangeResourceAmount(ResourceType resourceType, double change);
 	}
 }
